Move classes from threading_control.h to separate files
