+++ src/tbb/CMakeLists.txt
@@ -40,6 +40,7 @@ add_library(tbb
     task_dispatcher.cpp
     task_group_context.cpp
     thread_dispatcher.cpp
+    thread_request_serializer.cpp
     threading_control.cpp
     version.cpp
     queuing_rw_mutex.cpp)
